[refac] Refactor package structure to centralize auth and member entities under core
