--- edukit-core/src/main/java/com/edukit/auth/entity/ValidEmail.java
@@ -1,27 +1,0 @@
-package com.edukit.auth.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ValidEmail {
-
-    @Id
-    @Column(name = "valid_email_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String validEmail;
-
-    @Column(nullable = false)
-    private String educationOffice;
-}
