--- edukit-core/src/main/java/com/edukit/auth/enums/AuthCodeStatus.java
@@ -1,6 +1,0 @@
-package com.edukit.auth.enums;
-
-public enum AuthCodeStatus {
-    COMPLETE,
-    PENDING
-}
