--- edukit-core/src/main/java/com/edukit/auth/enums/AuthCodeType.java
@@ -1,6 +1,0 @@
-package com.edukit.auth.enums;
-
-public enum AuthCodeType {
-    TEACHER_VERIFICATION,
-    PASSWORD_RESET
-}
