--- edukit-core/src/main/java/com/edukit/auth/exception/AuthException.java
@@ -1,10 +1,0 @@
-package com.edukit.auth.exception;
-
-import com.edukit.common.exception.BusinessException;
-import com.edukit.common.exception.code.ErrorCode;
-
-public class AuthException extends BusinessException {
-    public AuthException(final ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
