--- edukit-core/src/main/java/com/edukit/common/config/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package com.edukit.common.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
