+++ edukit-core/src/main/java/com/edukit/core/auth/enums/AuthCodeStatus.java
@@ -1,0 +1,6 @@
+package com.edukit.core.auth.enums;
+
+public enum AuthCodeStatus {
+    COMPLETE,
+    PENDING
+}
