+++ edukit-core/src/main/java/com/edukit/core/auth/enums/AuthCodeType.java
@@ -1,0 +1,6 @@
+package com.edukit.core.auth.enums;
+
+public enum AuthCodeType {
+    TEACHER_VERIFICATION,
+    PASSWORD_RESET
+}
