+++ edukit-core/src/main/java/com/edukit/core/common/config/JpaAuditingConfig.java
@@ -1,0 +1,9 @@
+package com.edukit.core.common.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+}
