+++ edukit-core/src/main/java/com/edukit/core/member/entity/NicknameBannedWord.java
@@ -1,0 +1,25 @@
+package com.edukit.core.member.entity;
+
+import com.edukit.core.common.domain.BaseTimeEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class NicknameBannedWord extends BaseTimeEntity {
+
+    @Id
+    @Column(name = "nickname_banned_word_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String word;
+}
