+++ edukit-core/src/main/java/com/edukit/core/member/entity/Subject.java
@@ -1,0 +1,36 @@
+package com.edukit.core.member.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Subject {
+
+    @Id
+    @Column(name = "subject_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Builder(access = AccessLevel.PRIVATE)
+    private Subject(final String name) {
+        this.name = name;
+    }
+
+    public static Subject create(final String name) {
+        return Subject.builder()
+                .name(name)
+                .build();
+    }
+}
