+++ edukit-core/src/main/java/com/edukit/core/member/enums/MemberRole.java
@@ -1,0 +1,14 @@
+package com.edukit.core.member.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum MemberRole {
+    TEACHER("ROLE_TEACHER"),
+    PENDING_TEACHER("ROLE_PENDING_TEACHER"),
+    ADMIN("ROLE_ADMIN");
+
+    private final String name;
+}
