+++ edukit-core/src/main/java/com/edukit/core/member/exception/MemberException.java
@@ -1,0 +1,10 @@
+package com.edukit.core.member.exception;
+
+import com.edukit.common.exception.BusinessException;
+import com.edukit.common.exception.code.ErrorCode;
+
+public class MemberException extends BusinessException {
+    public MemberException(final ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
