+++ edukit-core/src/main/java/com/edukit/core/notice/entity/Notice.java
@@ -1,0 +1,58 @@
+package com.edukit.core.notice.entity;
+
+import com.edukit.core.common.domain.BaseTimeEntity;
+import com.edukit.core.notice.enums.NoticeCategory;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import java.time.LocalDateTime;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.SQLRestriction;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@SQLRestriction("is_deleted = false")
+public class Notice extends BaseTimeEntity {
+    @Id
+    @Column(name = "notice_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    private NoticeCategory category;
+
+    @Column(nullable = false)
+    private String title;
+
+    @Column(nullable = false, columnDefinition = "TEXT")
+    private String content;
+
+    @Column(nullable = false)
+    private boolean isDeleted = false;
+
+    private LocalDateTime deletedAt;
+
+    @Builder(access = AccessLevel.PRIVATE)
+    private Notice(final NoticeCategory category, final String title, final String content) {
+        this.category = category;
+        this.title = title;
+        this.content = content;
+    }
+
+    public static Notice create(final NoticeCategory category, final String title, final String content) {
+        return Notice.builder()
+                .category(category)
+                .title(title)
+                .content(content)
+                .build();
+    }
+}
