+++ edukit-core/src/main/java/com/edukit/core/notice/exception/NoticeException.java
@@ -1,0 +1,10 @@
+package com.edukit.core.notice.exception;
+
+import com.edukit.common.exception.BusinessException;
+import com.edukit.common.exception.code.ErrorCode;
+
+public class NoticeException extends BusinessException {
+    public NoticeException(final ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
