+++ edukit-core/src/main/java/com/edukit/core/post/enums/PostCategory.java
@@ -1,0 +1,4 @@
+package com.edukit.core.post.enums;
+
+public enum PostCategory {
+}
