+++ edukit-core/src/main/java/com/edukit/core/post/enums/PostFileCategory.java
@@ -1,0 +1,6 @@
+package com.edukit.core.post.enums;
+
+public enum PostFileCategory {
+    ORIGINAL,
+    PREVIEW
+}
