+++ edukit-core/src/main/java/com/edukit/core/post/exception/PostException.java
@@ -1,0 +1,10 @@
+package com.edukit.core.post.exception;
+
+import com.edukit.common.exception.BusinessException;
+import com.edukit.common.exception.code.ErrorCode;
+
+public class PostException extends BusinessException {
+    public PostException(final ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
