+++ edukit-core/src/main/java/com/edukit/core/studentrecord/enums/StudentRecordType.java
@@ -1,0 +1,9 @@
+package com.edukit.core.studentrecord.enums;
+
+public enum StudentRecordType {
+    SUBJECT,    // 세부능력 및 특기사항
+    BEHAVIOR,   // 행동특성 및 종합의견
+    CAREER,     // 창의적 체험활동 - 진로
+    FREE,       // 창의적 체험활동 - 자율
+    CLUB        // 창의적 체험활동 - 동아리
+}
