--- edukit-core/src/main/java/com/edukit/member/entity/NicknameBannedWord.java
@@ -1,25 +1,0 @@
-package com.edukit.member.entity;
-
-import com.edukit.common.domain.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class NicknameBannedWord extends BaseTimeEntity {
-
-    @Id
-    @Column(name = "nickname_banned_word_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String word;
-}
