--- edukit-core/src/main/java/com/edukit/member/enums/School.java
@@ -1,13 +1,0 @@
-package com.edukit.member.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum School {
-    MIDDLE_SCHOOL("middle"),
-    HIGH_SCHOOL("high");
-
-    private final String name;
-}
