--- edukit-core/src/main/java/com/edukit/member/exception/MemberException.java
@@ -1,10 +1,0 @@
-package com.edukit.member.exception;
-
-import com.edukit.common.exception.BusinessException;
-import com.edukit.common.exception.code.ErrorCode;
-
-public class MemberException extends BusinessException {
-    public MemberException(final ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
