--- edukit-core/src/main/java/com/edukit/notice/exception/NoticeException.java
@@ -1,10 +1,0 @@
-package com.edukit.notice.exception;
-
-import com.edukit.common.exception.BusinessException;
-import com.edukit.common.exception.code.ErrorCode;
-
-public class NoticeException extends BusinessException {
-    public NoticeException(final ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
