--- edukit-core/src/main/java/com/edukit/post/entity/PostFile.java
@@ -1,59 +1,0 @@
-package com.edukit.post.entity;
-
-import com.edukit.post.enums.PostFileCategory;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import java.time.LocalDateTime;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostFile {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "post_file_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id", nullable = false)
-    private Post post;
-
-    @Column(nullable = false)
-    @Enumerated(EnumType.STRING)
-    private PostFileCategory category;
-
-    @Column(nullable = false)
-    private String fileName;
-
-    @Column(length = 500, nullable = false)
-    private String filePath;
-
-    @Column(nullable = false)
-    private Long fileSize;
-
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
-
-    @Builder(access = AccessLevel.PRIVATE)
-    public PostFile(Post post, PostFileCategory category, String fileName, String filePath, Long fileSize, LocalDateTime createdAt) {
-        this.post = post;
-        this.category = category;
-        this.fileName = fileName;
-        this.filePath = filePath;
-        this.fileSize = fileSize;
-        this.createdAt = createdAt;
-    }
-}
