--- edukit-core/src/main/java/com/edukit/post/enums/PostCategory.java
@@ -1,4 +1,0 @@
-package com.edukit.post.enums;
-
-public enum PostCategory {
-}
