--- edukit-core/src/main/java/com/edukit/post/enums/PostFileCategory.java
@@ -1,6 +1,0 @@
-package com.edukit.post.enums;
-
-public enum PostFileCategory {
-    ORIGINAL,
-    PREVIEW
-}
