--- edukit-core/src/main/java/com/edukit/post/exception/PostException.java
@@ -1,10 +1,0 @@
-package com.edukit.post.exception;
-
-import com.edukit.common.exception.BusinessException;
-import com.edukit.common.exception.code.ErrorCode;
-
-public class PostException extends BusinessException {
-    public PostException(final ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
