--- edukit-core/src/main/java/com/edukit/studentrecord/enums/StudentRecordType.java
@@ -1,9 +1,0 @@
-package com.edukit.studentrecord.enums;
-
-public enum StudentRecordType {
-    SUBJECT,    // 세부능력 및 특기사항
-    BEHAVIOR,   // 행동특성 및 종합의견
-    CAREER,     // 창의적 체험활동 - 진로
-    FREE,       // 창의적 체험활동 - 자율
-    CLUB        // 창의적 체험활동 - 동아리
-}
