--- edukit-core/src/main/java/com/edukit/studentrecord/exception/StudentRecordException.java
@@ -1,10 +1,0 @@
-package com.edukit.studentrecord.exception;
-
-import com.edukit.common.exception.BusinessException;
-import com.edukit.common.exception.code.ErrorCode;
-
-public class StudentRecordException extends BusinessException {
-    public StudentRecordException(final ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
